+++ src/h/config_internal.h
@@ -1,0 +1,10 @@
+#ifndef HHC_CONFIG_INTERNAL_H
+#define HHC_CONFIG_INTERNAL_H
+
+#define MODE_INTERNAL_VSYNC                         0
+#define MODE_INTERNAL_LOAD_CHUNKS                   1
+#define MODE_INTERNAL_EXPORT_CHUNKS                 1
+#define MODE_INTERNAL_IMPORT_CHUNKS                 1
+#define MODE_INTERNAL_COLLIDE                       1
+
+#endif /* HHC_CONFIG_INTERNAL_H */